+++ common/src/main/java/net/xstopho/resourcelibrary/rendering/item/ItemModelRenderHelper.java
@@ -3,6 +3,9 @@ package net.xstopho.resourcelibrary.rendering.item;
 import net.minecraft.client.resources.model.ModelResourceLocation;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
+import net.xstopho.resourcelibrary.LibConstants;
+import net.xstopho.resourcelibrary.platform.IPlatformHelper;
+import net.xstopho.resourcelibrary.service.CoreServices;
 
 import java.util.HashMap;
 import java.util.LinkedHashMap;
@@ -12,6 +15,9 @@ public class ItemModelRenderHelper {
     private static final HashMap<Item, ModelResourceLocation> ITEM_MODELS = new LinkedHashMap<>();
 
     public static void registerItemModel(Item item, ResourceLocation itemTexturePath) {
+        if (CoreServices.PLATFORM == IPlatformHelper.Platforms.FORGE) {
+            LibConstants.LOG.error("Registering Custom Item Model isn't working in Forge");
+        }
         ITEM_MODELS.putIfAbsent(item, location(itemTexturePath));
     }