+++ neoforge/src/main/java/net/xstopho/resourcelibrary/test/LibraryTest.java
@@ -1,0 +1,10 @@
+package net.xstopho.resourcelibrary.test;
+
+public class LibraryTest {
+
+    public static void init() {
+        TestBlocks.init();
+        TestItems.init();
+        TestCreativeTab.init();
+    }
+}